Sprint semi fonctiennel du projet V2.00

+++ src/main/java/com/example/sallereservation/controller/DeleteUserServlet.java
@@ -1,4 +1,6 @@
 package com.example.sallereservation.controller;
+
+import com.example.sallereservation.model.User;
 import com.example.sallereservation.service.UserService;
 
 import javax.servlet.ServletException;
@@ -14,10 +16,16 @@ public class DeleteUserServlet extends HttpServlet {
 
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+        User loggedUser = (User) request.getSession().getAttribute("user");
+        if (loggedUser == null || !loggedUser.getRole().equals("admin")) {
+            response.sendRedirect(request.getContextPath() + "/accessDenied");
+            return;
+        }
+
         int userId = Integer.parseInt(request.getParameter("id"));
         userService.deleteUser(userService.getUserById(userId));
 
-        request.getSession().setAttribute("message", "User deleted successfully");
+        request.getSession().setAttribute("message", "Utilisateur supprimé avec succès !");
         response.sendRedirect("dashboard");
     }
 }
